BACK TO NATIVE UART
authored by Daniel Maslowski <info@orangecms.org>

+++ riscv64/src/memory.rs
@@ -2,6 +2,6 @@ use crate::platform::PHYSICAL_MEMORY_OFFSET;
 
 /// Convert physical address to virtual address
 #[inline]
-pub const fn phys_to_virt(paddr: usize) -> usize {
+pub const fn phys_to_virt(paddr: u64) -> u64 {
     PHYSICAL_MEMORY_OFFSET + paddr
 }
