+++ riscv64/src/platform/virt/devcons.rs
@@ -16,7 +16,7 @@ pub fn init(dt: &DeviceTree) {
         .unwrap();
 
     Console::new(|| {
-        let mut uart = Uart16550::new(phys_to_virt(ns16550a_reg.addr as usize) as usize);
+        let mut uart = Uart16550::new(phys_to_virt(ns16550a_reg.addr));
         uart.init(115_200);
 
         static mut UART: MaybeUninit<Uart16550> = MaybeUninit::uninit();
